Mock db refactoring
