--- common/db/mockDataStore.ts
@@ -1,6 +1,0 @@
-// TODO: add comment
-const JSONdb = require('../lib/Simple-JSONdb');
-
-const dataStore = new JSONdb('../storage.json');
-
-export default dataStore;
