+++ oauth-connect/pages/_app.tsx
@@ -3,5 +3,5 @@ import type { AppProps } from 'next/app';
 import '../../common/styles/globals.css';
 
 export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-};
+  return <Component {...pageProps} />;
+}
